requirements.txt
PyQt5~=5.15.4
mysql-connector-python~=8.0.28
environs~=9.3.5